Added linear_utree, a compact utree stored in linearized form
The linear_utree stores a compact utree in linearized form, like
a simple sorted array but with sparse elements. The idea was to
take advantage of the narrowing nature of binary search to avoid
cache misses when navigating the tree.

The REALLY interesting result was that this idea was horribly flawed
and the resulting performance of the linear_utree is already showing
to be much worse than an equivalent compact_utree.

The intention is to continue with the linear form of these trees
just to see how they compare with later optimizations added in case
there are additional surprises.