Modified shplot to max out at 80 columns
Before, large terminals => large plots => not easy to read plots.

Updated to not exceed 80 columns by default. Additionally, height
is not longer based on terminal height but on a ratio of width.
Most scripts will likely add output after the plot, so shrinking
based on terminal height will likely not be that useful.

Suggested by @kegilbert