Added inplace rebalancing for compact trees
Inplace rebalancing is accomplished by performing an in-order
traversal of a theoretical perfect tree. An in-order traversal
of the existing tree is performed simultaneously, and each element
is moved over in order.

Since both of these traversals occur in the same region of memory,
there is potential for the elements to become corrupted. To avoid
this, the tree is first transformed into the worst possible tree,
from the right to the left, then transformed into the perfect tree,
from the left to the right.

Comparison with 16ke compact_utree:

                        L1 misses   L2 misses   branch mispredicts
inplace rebalance:      8.72m/e     114mm/e     30.0m/e
naive rebalance:        15.2m/e     133mm/e     21.1m/e