Moved existance information into parent nodes
Instead of storing the existance of a node as a flag, the parent stores
the existance of both its left and right child. The three flags
associated with a node still fit in a word, so there should be no space
increase given reasonable padding.

The benefit of this approach is that fewer nodes need to be touched in
a full tree. This results in better cache usage and potential for lazily
allocated pages in sparse trees.

Comparison of 16ke compact_utree:

                        L1 misses   L2 misses   heap usage
parent exist flags:     6.50m/e     111mm/e     36B/e
leaf exist flags:       8.72m/e     133mm/e     36B/e